The gradient matrix of a_fun.

The gradient is taken with respect to the item parameters and organized to be conformable with Matrix::bdiag(mle$var.cov). When evaluating the gradient under the null hypothesis of no DIF, the optional argument theta can be provided. It replaces the item-specific values of a_fun in the gradient computation.

grad_a(mle, theta = NULL, log = F)

Arguments

mle

the output of get_model_parms

theta

(optional) the scaling parameter. Replaces item-specific values of alpha if provided.

log

logical: return of log(a2/a1)?

Value

A matrix in which the columns are the gradient vectors of a_fun, for each item.

See also